Reorganization of maximum-likelihood refinement
The old MaximumLikelihoodRefinementEngine has been renamed MLWithModelErrorsRefinementEngine.
Its likelihood function has as its major source of uncertainties an assumed error distribution
of the model parameters (positions and ADPs), which is appropriate for estimating phases
but not for structure refinement. A new class MLRefinementEngine implements a likelihood
function whose only uncertainties come from 1) the experimental error on the structure factor
amplitudes and 2) atoms not contained in the model.