ReflectionSet: new compact storage mode
A ReflectionSet used to keep a minimal list of reflections corresponding to an asymmetric unit,
but also keep a dictionary containing the Reflection objects for all symmetry-related reflections.
The latter are rarely used (they merely speed up the retrieval of the Reflection object
corresponding to a given set of Miller indices) but can occupy a lot of memory for high-resolution
data sets at high symmetry.

A second storage mode, named "compact" (and made the default) stores
Reflection objects only for an asymmetric unit. If a Reflection is requested for a set of
Miller indices outside this asymmetric unit, the symmetry-related Miller indices are generated
and looked up until the right one is found. While this operation is slow, it is not used
very frequently. Most applications iterate over the ReflectionSet in order to retrieve the
reflections in the asymmetric unit. This operation has the same speed in compact and
traditional storage mode.

As a result of this change, pickled ReflectionSet objects from older versions can no longer
be read. The version number was increased to indicate this incompatibility.