Unified atom iterator interface
The ReflectionData classes and the RefinementEngine classes now use the same atom iterator
interface. This implies a change in the interface for the ReflectionData classes that
previously did not expect the atom_id item. The class Crystal now includes a list of atoms
and can be used directly as the iterator argument wherever an atom iterator is required.