New mmCIF parser
The old mmCIF parser taken from PyMMLIB reads the whole file into a set of not very
efficient data structures, requiring hundreds of megabytes of memory for large structure
factor files. The new one is implemented as an iterator that stores nothing at all,
plus a high-level interface that uses callbacks to let the client store the data
as it is read directly into the data structures that will be used. It reduces
the memory footprint of mmCIF-reading scripts dramatically and usually makes them
faster as well.